# apps/ai/openapi.json
{
  "components": {
    "schemas": {
      "AssistantEvent": {
        "description": "One server-sent event of an answer.",
        "properties": {
          "event": {
            "discriminator": {
              "mapping": {
                "done": "#/components/schemas/DoneEvent",
                "error": "#/components/schemas/ErrorEvent",
                "sources": "#/components/schemas/SourcesEvent",
                "text": "#/components/schemas/TextEvent"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/TextEvent"
              },
              {
                "$ref": "#/components/schemas/SourcesEvent"
              },
              {
                "$ref": "#/components/schemas/DoneEvent"
              },
              {
                "$ref": "#/components/schemas/ErrorEvent"
              }
            ],
            "title": "Event"
          }
        },
        "required": [
          "event"
        ],
        "title": "AssistantEvent",
        "type": "object"
      },
      "AssistantRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AssistantRequest",
        "type": "object"
      },
      "DocumentCreate": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "description": "Plain text.",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "DocumentCreate",
        "type": "object"
      },
      "DocumentOut": {
        "properties": {
          "chunkCount": {
            "title": "Chunkcount",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdby"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "indexing",
              "ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "error",
          "chunkCount",
          "summary",
          "createdBy",
          "createdAt"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
      "DoneEvent": {
        "properties": {
          "type": {
            "const": "done",
            "default": "done",
            "title": "Type",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "usage"
        ],
        "title": "DoneEvent",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "BAD_REQUEST",
          "VALIDATION_FAILED",
          "UNAUTHENTICATED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "RATE_LIMITED",
          "INTERNAL",
          "SERVICE_UNAVAILABLE",
          "METHOD_NOT_SUPPORTED",
          "TIMEOUT",
          "PAYLOAD_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "IDEMPOTENCY_IN_PROGRESS",
          "FEATURE_DISABLED",
          "CLIENT_OUTDATED",
          "FRESH_SESSION_REQUIRED",
          "DESTINATION_NOT_ALLOWED",
          "RESPONSE_TOO_LARGE",
          "TOO_MANY_REDIRECTS",
          "UPSTREAM_UNAVAILABLE",
          "NO_ACTIVE_ORGANIZATION",
          "PHONE_CODE_INVALID",
          "PHONE_NUMBER_TAKEN",
          "ENTITLEMENT_REQUIRED",
          "ALREADY_SUBSCRIBED",
          "NO_SUBSCRIPTION",
          "FILE_NOT_FOUND",
          "FILE_TYPE_NOT_ALLOWED",
          "FILE_TOO_LARGE",
          "FILE_SIZE_MISMATCH",
          "FILE_INFECTED",
          "FILE_UNREADABLE",
          "FILE_NOT_UPLOADED",
          "FILE_NOT_READY",
          "UNSUBSCRIBE_LINK_INVALID",
          "INVALID_SIGNATURE",
          "WEBHOOK_ENDPOINT_NOT_FOUND",
          "WEBHOOK_DELIVERY_NOT_FOUND",
          "WEBHOOK_URL_NOT_ALLOWED",
          "WEBHOOK_ENDPOINT_LIMIT",
          "APP_NOT_FOUND",
          "API_KEY_NOT_FOUND",
          "API_KEY_LIMIT_REACHED",
          "API_KEY_SCOPE_MISSING",
          "DOCUMENT_NOT_FOUND",
          "DOCUMENT_INDEXING_FAILED",
          "AI_BUDGET_EXCEEDED",
          "AI_RUN_LIMIT",
          "TODO_NOT_FOUND",
          "TODO_VERSION_CONFLICT"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorData": {
        "additionalProperties": false,
        "properties": {
          "issues": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issues"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            },
            "title": "Params",
            "type": "object"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid"
          }
        },
        "required": [
          "params"
        ],
        "title": "ErrorData",
        "type": "object"
      },
      "ErrorEvent": {
        "properties": {
          "code": {
            "enum": [
              "AI_RUN_LIMIT",
              "UPSTREAM_UNAVAILABLE"
            ],
            "title": "Code",
            "type": "string"
          },
          "type": {
            "const": "error",
            "default": "error",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
      "ErrorIssue": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "path": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            },
            "title": "Path",
            "type": "array"
          }
        },
        "required": [
          "path",
          "code"
        ],
        "title": "ErrorIssue",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorData"
          },
          "defined": {
            "title": "Defined",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "defined",
          "code",
          "status",
          "message",
          "data"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "SentimentRequest": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "description": "Text to classify.",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "SentimentRequest",
        "type": "object"
      },
      "SentimentResponse": {
        "properties": {
          "label": {
            "enum": [
              "positive",
              "negative",
              "neutral"
            ],
            "title": "Label",
            "type": "string"
          },
          "model": {
            "description": "Identifier of the model that produced the prediction.",
            "title": "Model",
            "type": "string"
          },
          "score": {
            "description": "Confidence of the predicted label.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "label",
          "score",
          "model"
        ],
        "title": "SentimentResponse",
        "type": "object"
      },
      "Source": {
        "properties": {
          "documentId": {
            "format": "uuid",
            "title": "Documentid",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "documentId",
          "title"
        ],
        "title": "Source",
        "type": "object"
      },
      "SourcesEvent": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          },
          "type": {
            "const": "sources",
            "default": "sources",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "sources"
        ],
        "title": "SourcesEvent",
        "type": "object"
      },
      "TextEvent": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "const": "text",
            "default": "text",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TextEvent",
        "type": "object"
      },
      "Usage": {
        "properties": {
          "inputTokens": {
            "title": "Inputtokens",
            "type": "integer"
          },
          "outputTokens": {
            "title": "Outputtokens",
            "type": "integer"
          }
        },
        "required": [
          "inputTokens",
          "outputTokens"
        ],
        "title": "Usage",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ai",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health/dependencies": {
      "get": {
        "description": "Postgres and Redis answer: for dashboards and start-up scripts.",
        "operationId": "dependencies",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An error."
          }
        },
        "summary": "Dependencies"
      }
    },
    "/health/live": {
      "get": {
        "operationId": "live",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An error."
          }
        },
        "summary": "Live"
      }
    },
    "/health/ready": {
      "get": {
        "description": "Serving requests. Not the dependencies: they're shared, so an outage of one would\ntake every pod out of rotation at once (see /health/dependencies).",
        "operationId": "ready",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An error."
          }
        },
        "summary": "Ready"
      }
    },
    "/v1/assistant/answers": {
      "post": {
        "operationId": "answer",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantEvent"
                }
              },
              "text/event-stream": {}
            },
            "description": "Server-sent events, each `data:` line one AssistantEvent."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An error."
          }
        },
        "summary": "Answer"
      }
    },
    "/v1/documents": {
      "get": {
        "operationId": "listDocuments",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response Listdocuments",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An error."
          }
        },
        "summary": "List Documents"
      },
      "post": {
        "operationId": "createDocument",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An error."
          }
        },
        "summary": "Create Document"
      }
    },
    "/v1/documents/{document_id}": {
      "delete": {
        "operationId": "deleteDocument",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An error."
          }
        },
        "summary": "Delete Document"
      }
    },
    "/v1/sentiment": {
      "post": {
        "operationId": "sentiment",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SentimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentimentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An error."
          }
        },
        "summary": "Sentiment"
      }
    }
  }
}